finish 1.21.5 port
loom 1.7 can no longer be used, so this commit also updates it to 1.10.5, bumping the required gradle version to 8.12.

not sure if all the .get() calls are good code hygiene. my intuition says these need some backup checks.... whatever, i'd like to get it running first and proper second

i was hoping this would work on 1.21.6 but unfortunately there's more changes in there and a mixin's failing. bumping to 1.21.6 next, then i'll rebase