most of a 1.21.6 port
writeview and readview are cool but unfortunately we need to read in a mixin to readview. this creates a cascade of utter bs that i patched over until it built, except it very much does not work. i'm not qualified to rewrite the entire player/server summary api to match this so this is now a sisby problem